use TensorRT as a compiler in CUDA
