+++ include/onnxruntime/core/providers/cuda/cuda_provider_factory.h
@@ -10,7 +10,7 @@ extern "C" {
 /**
  * \param device_id cuda device id, starts from zero.
  */
-ORT_API_STATUS(OrtSessionOptionsAppendExecutionProvider_CUDA, _In_ OrtSessionOptions* options, int device_id);
+ORT_API_STATUS(OrtSessionOptionsAppendExecutionProvider_CUDA, _In_ OrtSessionOptions* options, int device_id, bool enable_compiler = false);
 
 #ifdef __cplusplus
 }
