--- onnxruntime/core/providers/tensorrt/symbols.txt
@@ -1,1 +1,0 @@
-OrtSessionOptionsAppendExecutionProvider_Tensorrt
