+++ onnxruntime/core/providers/tensorrt/tensorrt_compiler.h
@@ -2,15 +2,16 @@
 // Licensed under the MIT License.
 
 #pragma once
-#include <ctime>
 #include "core/common/logging/logging.h"
-#include "core/framework/op_kernel.h"
 #include "NvInfer.h"
 #include "NvOnnxParser.h"
+#include "core/framework/compute_capability.h"
+#include "core/graph/graph_viewer.h"
 #include "core/platform/ort_mutex.h"
+#include "core/framework/op_kernel.h"
 
 namespace onnxruntime {
-
+struct NodeComputeInfo;
 class TensorrtLogger : public nvinfer1::ILogger {
   nvinfer1::ILogger::Severity verbosity_;
 
@@ -30,10 +31,19 @@ class TensorrtLogger : public nvinfer1::ILogger {
   }
 };
 
-// Information needed to construct trt execution providers.
-struct TensorrtExecutionProviderInfo {
-  int device_id{0};
+struct InferDeleter {
+  template <typename T>
+  void operator()(T* obj) const {
+    if (obj) {
+      obj->destroy();
+    }
+  }
 };
+
+template <typename T>
+using unique_pointer = std::unique_ptr<T, InferDeleter>;
+
+std::vector<std::unique_ptr<ComputeCapability>> GetTensorrtCapability(const onnxruntime::GraphViewer& graph);
 
 // Information to construct kernel function state.
 struct TensorrtFuncState {
@@ -47,68 +57,27 @@ struct TensorrtFuncState {
   std::vector<std::vector<int>> output_info;
   std::vector<std::vector<int64_t>> output_shapes;
   OrtMutex* tensorrt_mu_ptr = nullptr;
+  cudaStream_t stream;
 };
 
-// Logical device representation.
-class TensorrtExecutionProvider : public IExecutionProvider {
+class TensorrtCompiler {
  public:
-  TensorrtExecutionProvider();
-  virtual ~TensorrtExecutionProvider();
-
-  std::vector<std::unique_ptr<ComputeCapability>>
-  GetCapability(const onnxruntime::GraphViewer& graph,
-                const std::vector<const KernelRegistry*>& /*kernel_registries*/) const override;
-
-  common::Status Compile(const std::vector<onnxruntime::Node*>& fused_nodes,
-                         std::vector<NodeComputeInfo>& node_compute_funcs) override;
-
-  void SetMaxBatchSize(const int batch_size) {
-    max_batch_size_ = batch_size;
-  }
-
-  void SetMaxWorkspaceSize(const size_t workspace_size) {
-    max_workspace_size_ = workspace_size;
-  }
+  TensorrtCompiler();
+  common::Status Compile(const onnxruntime::Node* fused_nodes, cudaStream_t stream,
+                         NodeComputeInfo& node_compute_funcs);
 
  private:
-  int max_batch_size_ = 1;
+  size_t max_batch_size_ = 1;
   size_t max_workspace_size_ = 1 << 30;  // 1GB
-  int max_parser_iterations_ = 6;
-
-  struct InferDeleter {
-    template <typename T>
-    void operator()(T* obj) const {
-      if (obj) {
-        obj->destroy();
-      }
-    }
-  };
-
-  template <typename T>
-  using unique_pointer = std::unique_ptr<T, InferDeleter>;
 
   OrtMutex tensorrt_mu_;
   int device_id_;
+
   std::unordered_map<std::string, unique_pointer<nvonnxparser::IParser>> parsers_;
   std::unordered_map<std::string, unique_pointer<nvinfer1::ICudaEngine>> engines_;
   std::unordered_map<std::string, unique_pointer<nvinfer1::IExecutionContext>> contexts_;
   std::unordered_map<std::string, std::vector<std::vector<int>>> input_info_;
   std::unordered_map<std::string, std::vector<std::vector<int>>> output_info_;
   std::unordered_map<std::string, std::vector<std::vector<int64_t>>> output_shapes_;
-
-  /**Get IndexedSubGraph based on node list of the subgraph*/
-  std::unique_ptr<IndexedSubGraph> GetSubGraph(SubGraph_t graph_nodes_index, int& kernels_index,
-                                               const onnxruntime::GraphViewer& graph) const;
-
-  /**
-  Get TensorRT supported node lists by calling Onnx-TensorRT parser recursively. Since each time the parser
-  can only detect first unsupported node failure, it needs to wait for Onnxruntime to partition the graph
-  and then detect next failure again. If there are too many iterations, which means many nodes in the graph
-  are not supported by TensorRT, the process will be terminated and the whole graph is simply assigned to
-  other execution provider.
-  */
-  SubGraphCollection_t GetSupportedList(SubGraphCollection_t supported_nodes_list, int iterations, const int max_iterations,
-                                        const onnxruntime::GraphViewer& graph, bool* early_termination) const;
 };
-
 }  // namespace onnxruntime
