+++ onnxruntime/test/providers/provider_test_utils.cc
@@ -460,8 +460,6 @@ void OpTester::Run(ExpectResult expect_result,
           execution_provider = DefaultNupharExecutionProvider();
         else if (provider_type == onnxruntime::kBrainSliceExecutionProvider)
           execution_provider = DefaultBrainSliceExecutionProvider();
-        else if (provider_type == onnxruntime::kTensorrtExecutionProvider)
-          execution_provider = DefaultTensorrtExecutionProvider();
         else if (provider_type == onnxruntime::kOpenVINOExecutionProvider)
           execution_provider = DefaultOpenVINOExecutionProvider();
         else if (provider_type == onnxruntime::kNnapiExecutionProvider)
